Merge from homebrew, fix mutt_dotlock group
Styled the formula to better match that of the main homebrew branch.

Also, needed to change some things to support having mutt_dotlock not be
SGID mail, but made it SGID admin.  Not sure if this is a good thing to
have or if it is better to just disable it.